Updated contract macros. Now it copy implementation from trait definition and paste implementation in impl section directly. But it means that we copy implementation of the child contract directly into current contract. It requires us to make all parts of implementation public and import them like `contract::traits::*;`. But, it simplifies the code better.
