+++ contracts/access/access-control/lib.rs
@@ -1,5 +1,4 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
-pub mod impls;
 pub mod traits;
 mod tests;
