+++ contracts/brush/lib.rs
@@ -4,6 +4,6 @@ pub mod traits;
 pub use proc_macros::{
     contract,
     trait_definition,
-    internal_trait_definition,
+    storage_trait,
     modifiers,
 };
