+++ contracts/brush/proc_macros/lib.rs
@@ -4,6 +4,7 @@ mod contract;
 
 use quote::{
     quote,
+    ToTokens,
 };
 use syn::{
     ItemTrait,
@@ -24,33 +25,64 @@ pub fn contract(_attrs: TokenStream, ink_module: TokenStream) -> TokenStream {
 
 #[proc_macro_attribute]
 pub fn trait_definition(_attrs: TokenStream, _input: TokenStream) -> TokenStream {
-    let input: TokenStream2 = _input.clone().into();
     let attrs: TokenStream2 = _attrs.into();
-    let trait_item = parse_macro_input!(_input as ItemTrait);
+    let mut trait_item = parse_macro_input!(_input as ItemTrait);
 
+    // Save trait definition with generics and default methods to metadata.
     let locked_file = internal::get_locked_file();
-    let mut metadata = internal::load_metadata(&locked_file);
-    internal::put_trait(&mut metadata.external_traits, trait_item);
-    internal::save_metadata_and_unlock(locked_file, metadata);
+    let mut metadata = internal::Metadata::load(&locked_file);
+    metadata.external_traits.insert(
+        trait_item.ident.to_string(), trait_item.clone().into_token_stream().to_string());
+    metadata.save_and_unlock(locked_file);
+
+    // ink! doesn't support super trait and default functions, so we need to clean up it
+    // Remove super trait
+    trait_item.colon_token = None;
+    trait_item.supertraits.clear();
+    // Remove each default block and add semi colon at the end
+    trait_item.items
+        .iter_mut()
+        .for_each(|item| {
+            if let syn::TraitItem::Method(method) = item {
+                method.default = None;
+                method.semi_token = Some(syn::token::Semi::default());
+            }
+        });
+
+    // Remove all not ink functions
+    trait_item.items = trait_item.items.clone()
+        .into_iter()
+        .filter_map(|item|
+            if let syn::TraitItem::Method(method) = &item {
+                if internal::is_attr(&method.attrs, "ink") {
+                    Some(item)
+                } else {
+                    None
+                }
+            } else {
+                Some(item)
+            })
+        .collect();
 
     let code = quote! {
         #attrs
         #[ink_lang::trait_definition]
-        #input
+        #trait_item
     };
     code.into()
 }
 
 #[proc_macro_attribute]
-pub fn internal_trait_definition(_attrs: TokenStream, _input: TokenStream) -> TokenStream {
+pub fn storage_trait(_attrs: TokenStream, _input: TokenStream) -> TokenStream {
     let input: TokenStream2 = _input.clone().into();
     let attrs: TokenStream2 = _attrs.into();
     let trait_item = parse_macro_input!(_input as ItemTrait);
 
     let locked_file = internal::get_locked_file();
-    let mut metadata = internal::load_metadata(&locked_file);
-    internal::put_trait(&mut metadata.internal_traits, trait_item);
-    internal::save_metadata_and_unlock(locked_file, metadata);
+    let mut metadata = internal::Metadata::load(&locked_file);
+    metadata.storage_traits.insert(
+        trait_item.ident.to_string(), trait_item.into_token_stream().to_string());
+    metadata.save_and_unlock(locked_file);
 
     let code = quote! {
         #attrs
