+++ contracts/token/psp20/lib.rs
@@ -1,6 +1,5 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
-pub mod impls;
 pub mod traits;
 pub mod stub;
 mod tests;
