+++ contracts/token/psp20/tests.rs
@@ -2,11 +2,9 @@
 #[brush::contract]
 mod tests {
     /// Imports all the definitions from the outer scope so we can use them here.
-    use crate::impls::{PSP20, PSP20Storage, StorageHashMap, Lazy};
-    use crate::traits::{IPSP20};
-    use ink_prelude::{string::{String}};
+    use crate::traits::*;
     use ink_lang as ink;
-    use ink::{Env, EmitEvent};
+    use ink::{EmitEvent};
     use ink_env::{
         hash::{
             Blake2x256,
@@ -40,12 +38,12 @@ mod tests {
 
     /// A simple PSP-20 contract.
     #[ink(storage)]
-    #[derive(Default, PSP20Storage, IPSP20)]
+    #[derive(Default, PSP20Storage)]
     pub struct PSP20Struct {}
     type Event = <PSP20Struct as ::ink_lang::BaseEvent>::Type;
 
-    impl PSP20 for PSP20Struct {
-        fn emit_transfer_event(&self, _from: Option<AccountId>, _to: Option<AccountId>, _amount: Balance) {
+    impl IPSP20 for PSP20Struct {
+        fn _emit_transfer_event(&self, _from: Option<AccountId>, _to: Option<AccountId>, _amount: Balance) {
             self.env().emit_event(Transfer {
                 from: _from,
                 to: _to,
@@ -53,7 +51,7 @@ mod tests {
             });
         }
 
-        fn emit_approval_event(&self, _owner: AccountId, _spender: AccountId, _amount: Balance) {
+        fn _emit_approval_event(&self, _owner: AccountId, _spender: AccountId, _amount: Balance) {
             self.env().emit_event(Approval {
                 owner: _owner,
                 spender: _spender,
@@ -63,14 +61,10 @@ mod tests {
     }
 
     impl PSP20Struct {
-        pub fn new(_total_supply: Balance) -> impl PSP20 {
-            Self::constructor(_total_supply)
-        }
-
         #[ink(constructor)]
-        pub fn constructor(_total_supply: Balance) -> Self {
+        pub fn new(_total_supply: Balance) -> Self {
             let mut instance = Self::default();
-            instance.mint(instance.env().caller(), _total_supply);
+            instance._mint(instance.env().caller(), _total_supply);
             instance
         }
     }
