+++ contracts/token/psp721/lib.rs
@@ -1,6 +1,5 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod traits;
-pub mod impls;
 pub mod stub;
 mod tests;
