+++ examples/psp20/lib.rs
@@ -2,19 +2,15 @@
 
 #[brush::contract]
 pub mod my_psp20 {
-    use psp20::{
-        traits::{ IPSP20, PSP20Error },
-        impls::{ PSP20Storage, PSP20, StorageHashMap, Lazy, String },
-    };
+    use psp20::traits::*;
 
     #[ink(storage)]
-    #[derive(Default, PSP20Storage, IPSP20)]
+    #[derive(Default, PSP20Storage)]
     pub struct MyPSP20 {
         // fields for hater logic
         hated_account: AccountId,
     }
-
-    impl PSP20 for MyPSP20 {
+    impl IPSP20 for MyPSP20 {
         // Let's override method to reject transactions to bad account
         fn _before_token_transfer(&mut self, _from: AccountId, _to: AccountId, _amount: Balance) {
             assert!(_to != self.hated_account, "{}", PSP20Error::Unknown("I hate this account!").as_ref());
@@ -27,8 +23,8 @@ pub mod my_psp20 {
             let mut instance = Self::default();
             *instance._name_mut() = Lazy::new(name);
             *instance._symbol_mut() = Lazy::new(symbol);
-            instance.set_decimals(decimal);
-            instance.mint(instance.env().caller(), _total_supply);
+            *instance._decimals_mut() = Lazy::new(decimal);
+            instance._mint(instance.env().caller(), _total_supply);
             instance
         }
